+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   partial equality behaviour of floats.
 - `FloatEq` now more specifically means equality based on a structurally defined
   epsilon type. See the notes on `FloatEqAll` above.
+- `FloatDiff` is much more rigorously defined, in particular `ulps_diff` now
+  returns `Option<T>`, see the API documentation for details.
 - `FloatEq`'s `DiffEpsilon` is now `Epsilon`, and `UlpsDiffEpsilon` is now
   `UlpsEpsilon`. This reduces visual noise around the usage and harmonizes the
   naming with `FloatEqDebug`'s associated types.